// src/modules/llm/dtos/llm.dto.ts
import {
  IAISenderObject,
  TAISenderTagObject,
} from '../../sender/dtos/sender.dto';

export interface IAITagReport {
  confidence: number;
  description?: string | null;
  reason: string;
}

export interface ISentimentReport {
  overall: 'positive' | 'negative' | 'neutral';
  score: number; // Range: -1 to 1
  emotions: {
    happiness: number; // Range: 0 to 1
    sadness: number; // Range: 0 to 1
    anger: number; // Range: 0 to 1
    fear: number; // Range: 0 to 1
    surprise: number; // Range: 0 to 1
  };
}

export interface IAITagReportObject {
  purchase: IAITagReport;
  payment: IAITagReport;
  inquiry: IAITagReport;
  complaint: IAITagReport;
  newsletter: IAITagReport;
  subscription: IAITagReport;
  advertisement: IAITagReport;
  supportRequest: IAITagReport;
  feedback: IAITagReport;
  urgent: IAITagReport;
  followUp: IAITagReport;
  internal: IAITagReport;
  external: IAITagReport;
  promotion: IAITagReport;
  event: IAITagReport;
  spam: IAITagReport;
  other: IAITagReport;
}

export interface IAIDomainObject {
  name: string;
  website: string;
  description: string;
  summary: string;
  tags: TAISenderTagObject;
}

export interface ILLMResponse {
  summary: string;
  description: string;
  sentiment: ISentimentReport;
  messageTags: IAITagReportObject;
  domain: IAIDomainObject;
  sender: IAISenderObject;
}